requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.11
djangorestframework==3.16.1
dotenv==0.9.9
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
idna==3.11
numpy==2.2.6
packaging==26.0
pandas==2.3.3
polyline==2.0.4
pyogrio==0.12.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.1
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
